Refactor: consolidate shared-expert TP module
- Move shared-expert TP compute, communication, and validation into the canonical expert_shared module
- Keep shared weight generation beside the implementation that consumes it
- Update Decode MoE to import the consolidated shared-expert entry
